Smoothing for heatmaps and 2d graphs
Summary:
Also efficiency and power insights.
Known issue: when we smooth the heatmap we're skipping lines where there is no data at all. I'll fix this in a future patch.

Test Plan: Go to http://localhost:5000/labs/car-behavior/index.html?demo and note that the heatmaps and 2d graphs are smoother. Also note that all of the "insight" text is valid.

Reviewers: ruthie

Reviewed By: ruthie

Subscribers: jenkins, dhruv

Differential Revision: https://phabricator.automatic.co/D14900