Group heatmaps with basic vehicle picker
Test Plan:
Log in, go to http://localhost:5000/labs/car-behavior/index.html and see the new group heatmap at the bottom of the page (takes a few seconds to load). Then see the "compare with" menu. Select the "other vehicles" item off of that menu and see a dialog come up. Select an option from the dialog and hit the "select" button; see the new group heatmap load.

For instructions on logging in, see:

https://phabricator.automatic.co/w/projects/dashboard/labs/

... and grab the token from here:

https://automatic.slack.com/archives/data-science-heatmaps/p1464984395000148

Reviewers: ruthie

Reviewed By: ruthie

Subscribers: jenkins, dhruv

Differential Revision: https://phabricator.automatic.co/D14991