Set ranges for heatmaps and 2D graphs
Test Plan:
Log in and go to http://localhost:5000/labs/car-behavior/index.html and look at the various line graphs and heatmaps to see that they all have the same x-axis range. The heatmaps also have the same y-axis range as each other, but the graphs just follow the data on the y-axis.

Of course we don't have ticks in yet, so you can't see the numbers, but a sure sign is when you see a heatmap with a bunch of blank space on one side, as is the case with Jerry's 2013 Honda Fit.

For instructions on logging in, see:

https://phabricator.automatic.co/w/projects/dashboard/labs/

... and grab the token from here:

https://automatic.slack.com/archives/data-science-heatmaps/p1464984395000148

Reviewers: yanglin, ruthie

Reviewed By: yanglin

Subscribers: jenkins, dhruv

Differential Revision: https://phabricator.automatic.co/D15202