Using real vehicle picker data; ranges and labels in all graphs
Summary:
Also a couple of data changes:
* Cutting off MAF samples under a certain threshold
* Cutting off power samples in 2D graph where velocity is <= 0

As part of adding labels to the heatmap, I switched the heatmap over to SVG. It now takes around 100ms to draw one (instead of 10ms for Canvas) but it's probably worth it to keep things consistent.

To do in a future revision: make TwoDGraph inherit from SvgBase.

Test Plan:
Go to http://localhost:5000/labs/car-behavior/index.html and log in. Note that both line graphs and heatmaps now have labels and ranges displayed.

Select "other vehicles" from the "compare with" menu and select things in the dialog; notice that sometimes new options appear.

Reviewers: ruthie, yanglin

Reviewed By: yanglin

Subscribers: jenkins, dhruv

Differential Revision: https://phabricator.automatic.co/D15275