Various refinements for car behavior project
Summary:
* Group insights
* Renamed files because of deployment issue, see https://automatic.slack.com/archives/data-science-heatmaps/p1467835302000058
* Fixed display glitch with 2d graph labels when all data is 0
* Flavor text for comparisons

Test Plan:
Go to http://localhost:5000/labs/car-behavior/index.html and observe the above changes.

For the deployment fix, we'll just have to deploy and see if it worked. Note that @yanglin is planning on making a more robust fix to the deployment mechanism but has recommended the hack of renaming the files in the meantime.

Reviewers: ruthie

Reviewed By: ruthie

Subscribers: jenkins, yanglin, dhruv

Differential Revision: https://phabricator.automatic.co/D15872