Started car-health-monitor page
Summary:
Also, in car-behavior:
* Now possible to link to a specific tab
* Live resize
* Bigger gutter between heatmaps

Test Plan:
Go to http://localhost:5000/labs/car-health-monitor/ to see the new page; currently it's all dummy data and hipster ipsum. Will add graphs when the API is working. The details buttons should work, though. Make sure that the one for Power takes you to the right tab in the car-behavior page.

Go to http://localhost:5000/labs/car-behavior/ and resize the window, noting that the graphs and heatmaps resize properly. Also notice that the heatmaps are not quite as bunched together as before.

Reviewers: ruthie

Reviewed By: ruthie

Subscribers: jenkins, dhruv

Differential Revision: https://phabricator.automatic.co/D16135