Hard coding out of bounds weight to 1470
Summary: This is partially to deal with bad data we're getting from the car info service we're using called Chrome.

Test Plan: This bug was causing Jerry's Ford Mustang not to have Power graphs or heatmaps. Sign in as him and go to http://localhost:5000/labs/car-behavior/index.html and check out the power graphs.

Reviewers: ruthie

Reviewed By: ruthie

Subscribers: jenkins, dhruv

Differential Revision: https://phabricator.automatic.co/D16307