Update thresholds for showing graph data
Summary: Previously for Group graphs, the threshold was 300 points/seconds of data.  Now it is 1000.  For Individual driver graphs, the threshold was 30 points of data, now it is 60. We are generating these graphs over a larger set of data (more days analyzed) so the thresholds can be increased.

Test Plan: none

Reviewers: ian.gilman, ruthie

Reviewed By: ruthie

Subscribers: jenkins

Differential Revision: https://phabricator.automatic.co/D17139