Fix tagging conflict bug
Summary:
If a user tags a trip from some other client, and the state does not match the dashboard state, there will be a conflict.
Newton does not have a PATCH method for trip tagging, so on a conflict, we just fetch the current state, which should match what the user wants anyway.

Test Plan:
review, build, run
Will 404 or 409 errors generate sentry notifications?

Reviewers: pratik

Reviewed By: pratik

Subscribers: jenkins, ljuba, eric, crabclaw

Differential Revision: https://phabricator.automatic.co/D21585